+++ frontend/src/components/organisms/MatchesList/MatchesListProps.tsx
@@ -1,0 +1,20 @@
+export type MatchListProps = {
+    onClickTeam: () => void,
+    matchNumber: number | undefined,
+    level: number | undefined,
+    matchStatus: string | undefined
+    team1: {
+        readonly player1: string | undefined;
+        readonly player2: string | undefined;
+        readonly teamNumber: number | string | undefined;
+        readonly currentPlayerLevel: number | undefined;
+        readonly currentMatchNumber: number | undefined;
+    },
+    team2: {
+        readonly player1: string | undefined;
+        readonly player2: string | undefined;
+        readonly teamNumber: string | undefined;
+        readonly currentPlayerLevel: number | undefined;
+        readonly currentMatchNumber: number | undefined;
+    }
+};
